tweak: Move the entire system into the JSON file
Makes more sense than having to recompile just because you want to test
with a different set of networks, etc.

Also realised that `Public` and `Hidden` are outside of any network or
other permission - and moved them up a level in the struct.

Renamed a few things, added some more commentary.